Adding Sharpen Blade debuff
I'm intentionally not setting parent on this new debuff, because some people might want to see when an enemy warrior does a "go" (so they want to see Sharpen Blade on the nameplate) but don't necessarily want to see healing reduction debuffs on their frames. All the other healing reductions are toggle-able on their own, so I think this one should be as well.