Potential fix for the taint issue
Got this idea from another addon that might be a fix for potential taint issues when opening the editmode. I just added this junction 
function checkFrame(frame)
    if not issecurevariable(frame, "action") and not InCombatLockdown() then
        frame.action = nil
        frame:SetAttribute("action");
    end
end

and used it in every hooksecurfun call like 

  for _, frame in ipairs(ActionBarButtonEventsFrame.frames) do
	hooksecurefunc(frame, "UpdateAction", checkFrame);
  end

But this may need more testing because I think it create other issues. I use biggerbuff addon for druid to show more than 3 buffs. But it's only for showing more than 3 buffs on raid frame. Bigdebuff is more complicated so this is just a suggestion. Maybe it gives you an idea how to fix the issue.