Maybe a fix to the C stack overflow
Please, it would be great if someone could look over the commit. I'm rather new to the addon development. This change of line of code hasn't produced errors for me tonight. just moving these two lines out of the hooksecurefunc and it's only above the increase maxbuff function.

for _, frame in ipairs(ActionBarButtonEventsFrame.frames) do
	hooksecurefunc(frame, "UpdateAction", checkFrame);